table-order/backend/requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
sqlalchemy==2.0.40
pymysql==1.1.1
alembic==1.15.2
python-jose[cryptography]==3.4.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.1
pydantic==2.11.3
pydantic-settings==2.9.1
python-json-logger==3.3.0
sse-starlette==2.3.3
